Move ucd_genrator model one level up
