+++ java/jflex/ucd_generator/BUILD.bazel
@@ -10,9 +10,9 @@ java_library(
     deps = [
         "//java/jflex/ucd_generator/emitter/unicode_properties",
         "//java/jflex/ucd_generator/emitter/unicode_version",
+        "//java/jflex/ucd_generator/model",
         "//java/jflex/ucd_generator/scanner:exception",
         "//java/jflex/ucd_generator/scanner:ucd_scanners",
-        "//java/jflex/ucd_generator/scanner/model",
         "//java/jflex/ucd_generator/ucd",
         "//java/jflex/util/javac",
         "//java/jflex/velocity",
