+++ java/jflex/ucd_generator/scanner/AbstractEnumeratedPropertyFileScanner.java
@@ -2,7 +2,7 @@ package jflex.ucd_generator.scanner;
 
 import java.util.SortedSet;
 import java.util.TreeSet;
-import jflex.ucd_generator.scanner.model.UnicodeData;
+import jflex.ucd_generator.model.UnicodeData;
 import jflex.ucd_generator.ucd.NamedCodepointRange;
 
 /** Scans the common single-property Unicode.org data file format. */
