+++ java/jflex/ucd_generator/scanner/AbstractPropertyAliasesScanner.java
@@ -2,7 +2,7 @@ package jflex.ucd_generator.scanner;
 
 import java.util.HashSet;
 import java.util.Set;
-import jflex.ucd_generator.scanner.model.UnicodeData;
+import jflex.ucd_generator.model.UnicodeData;
 import jflex.ucd_generator.util.PropertyNameNormalizer;
 
 /** Scanner for {@code PropertyAliases(-X.X.X).txt}. */
