+++ java/jflex/ucd_generator/scanner/EnumeratedPropertyFileScanner.flex
@@ -1,6 +1,6 @@
 package jflex.ucd_generator.scanner;
 
-import jflex.ucd_generator.scanner.model.UnicodeData;
+import jflex.ucd_generator.model.UnicodeData;
 import jflex.ucd_generator.ucd.NamedCodepointRange;
 
 %%
