+++ java/jflex/ucd_generator/scanner/PropertyAliasesScanner.flex
@@ -1,6 +1,6 @@
 package jflex.ucd_generator.scanner;
 
-import jflex.ucd_generator.scanner.model.UnicodeData;
+import jflex.ucd_generator.model.UnicodeData;
 import jflex.ucd_generator.util.PropertyNameNormalizer;
 
 %%
