+++ javatests/jflex/ucd_generator/emitter/unicode_version/BUILD.bazel
@@ -5,7 +5,7 @@ java_test(
     deps = [
         "//java/jflex/testing/diff",
         "//java/jflex/ucd_generator/emitter/unicode_version",
-        "//java/jflex/ucd_generator/scanner/model",
+        "//java/jflex/ucd_generator/model",
         "//java/jflex/ucd_generator/ucd",
         "//third_party/com/google/truth",
     ],
